+++ frontend/src/common/enums/cars/filter-names.enum.ts
@@ -1,18 +1,25 @@
-enum FiltersNames {
-  REGION_ID = 'regionId',
+enum RangeNames {
+  YEAR = 'year',
+  PRICE = 'price',
+  RACE = 'race',
+}
+
+enum RangeValueNames {
   YEAR_START = 'yearStart',
   YEAR_END = 'yearEnd',
   PRICE_START = 'priceStart',
   PRICE_END = 'priceEnd',
   RACE_START = 'raceStart',
   RACE_END = 'raceEnd',
+}
+
+enum CheckListsNames {
+  REGION_ID = 'regionId',
   BODY_TYPE_ID = 'bodytypeId',
-  BRAND_ID = 'brandId',
-  MODEL_ID = 'modelId',
   COLOR_ID = 'colorId',
   TRANSMISSION_TYPE_ID = 'transmissionTypeId',
   FUEL_TYPE_ID = 'fueltypeId',
   DRIVETRAIN_ID = 'drivetrainId',
 }
 
-export { FiltersNames };
+export { RangeNames, RangeValueNames, CheckListsNames };
