+++ frontend/src/common/types/car-filter/autocomplete.type.ts
