+++ frontend/src/common/types/car-filter/brand-details.type.ts
@@ -2,4 +2,6 @@ export type BrandDetailsType = {
   id: string;
   brandId: string;
   modelId: string;
+  brandName: string;
+  modelName: string;
 };
