+++ frontend/src/common/types/car-filter/checkbox-list-data.type.ts
@@ -1,0 +1,6 @@
+import { CheckListsNames } from '../../enums/cars/filter-names.enum';
+
+export type CheckboxListDataType = {
+  filterName: CheckListsNames;
+  data: string[];
+};
