+++ frontend/src/common/types/car-filter/range-filters.type.ts
@@ -1,0 +1,16 @@
+type RangeFiltersType = {
+  year: {
+    yearStart: string;
+    yearEnd: string;
+  };
+  price: {
+    priceStart: string;
+    priceEnd: string;
+  };
+  race: {
+    raceStart: string;
+    raceEnd: string;
+  };
+};
+
+export type { RangeFiltersType };
