+++ frontend/src/common/types/car-filter/range-value.type.ts
@@ -1,0 +1,6 @@
+import { RangeNames } from '@common//enums/cars/filter-names.enum';
+
+export type RangeValueType = {
+  rangeName: RangeNames;
+  values: { [p: string]: string };
+};
