--- frontend/src/common/types/cars/checkbox-list-data.type.ts
@@ -1,6 +1,0 @@
-import { FiltersNames } from '../../enums/cars/filters-names.enum';
-
-export type CheckboxListDataType = {
-  filterName: FiltersNames;
-  data: string[];
-};
