--- frontend/src/common/types/cars/filters.type.ts
@@ -1,18 +1,0 @@
-type FiltersType = {
-  regionId: string;
-  yearStart: string;
-  yearEnd: string;
-  priceStart: string;
-  priceEnd: string;
-  raceStart: string;
-  raceEnd: string;
-  bodytypeId: string[];
-  brandId: string[];
-  modelId: string[];
-  colorId: string[];
-  transmissionTypeId: string[];
-  fueltypeId: string[];
-  drivetrainId: string[];
-};
-
-export type { FiltersType };
