--- frontend/src/common/types/cars/range-item.type.ts
@@ -1,6 +1,0 @@
-import { FiltersNames } from '../../enums/cars/filters-names.enum';
-
-export type RangeValueType = {
-  filterName: FiltersNames;
-  value: string;
-};
